requirements.txt
# Core Dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
